restructure settings page into logical submodule
